--- backend_draft/bin/.gitignore
@@ -1,1 +1,0 @@
-/backend_proj/
