+++ backend_draft/src/application/Meeting.java
