--- backend_draft/src/backend_proj/Location.java
@@ -1,13 +1,0 @@
-package backend_proj;
-
-import java.util.ArrayList;
-
-public class Location {
-
-	
-	private String department;
-	private int room;
-	private ArrayList<User> staff;
-	
-	
-}
