--- backend_draft/src/backend_proj/Meeting.java
@@ -1,10 +1,0 @@
-package backend_proj;
-
-import java.util.ArrayList;
-
-public class Meeting extends Booking {
-
-	
-	protected ArrayList<User> staff;
-	
-}
