--- backend_draft/src/backend_proj/Staff.java
@@ -1,5 +1,0 @@
-package backend_proj;
-
-public class Staff {
-
-}
