--- backend_draft/src/backend_proj/TestRecommendation.java
@@ -1,9 +1,0 @@
-package backend_proj;
-
-public class TestRecommendation {
-
-	
-	String testName;
-	String reason;
-	
-}
