--- backend_draft/src/backend_proj/User.java
@@ -1,11 +1,0 @@
-package backend_proj;
-
-public class User {
-
-	
-	protected String name;
-	protected String emailAddress;
-	protected String phoneNumber;
-	
-	
-}
