--- fictional-journey-master/README.md
@@ -1,2 +1,0 @@
-# fictional-journey
-A repository for the SENG 300 Group Project
