--- fictional-journey-master/backend_draft/src/application/Admin.java
@@ -1,28 +1,0 @@
-package application;
-
-import java.util.ArrayList;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-
-public class Admin extends User {
-
-    // default constructor
-    public Admin(String name, String email, String phone) {
-        super(name, email, phone);
-    }
-
-    // modify database
-    public void addDoctor(Database data, Doctor doctor) {
-        data.doctors.add(doctor);
-    }
-
-    public void addPatient(Database data, Patient patient) {
-        data.patients.add(patient);
-    }
-
-    public void addAdmin(Database data, Admin admin) {
-        data.administrators.add(admin);
-    }
-
-}
